Refactor any2mochi to use CLI parsing
