--- tools/any2mochi/config.go
@@ -1,5 +1,0 @@
-package any2mochi
-
-// UseLSP controls whether language servers are used for parsing.
-// It is disabled by default to avoid heavy dependencies during tests.
-var UseLSP = false
