+++ tools/any2mochi/server_stub.go
@@ -1,0 +1,10 @@
+package any2mochi
+
+// LanguageServer represents an external parser command.
+type LanguageServer struct {
+	Command string
+	Args    []string
+	LangID  string
+}
+
+var Servers = map[string]LanguageServer{}
